+++ blocktx/async/consumer.go
@@ -1,0 +1,5 @@
+package async
+
+type Consumer interface {
+	ConsumeTransactions() error
+}