+++ blocktx/interface.go
@@ -1,13 +1,7 @@
 package blocktx
 
-import "github.com/pkg/errors"
+import "errors"
 
 var (
-	// ErrBlockNotFound is returned when a block is not found.
-	ErrBlockNotFound = errors.New("block not found")
-
-	// ErrTransactionNotFound is returned when a transaction is not found.
-	ErrTransactionNotFound = errors.New("transaction not found")
-
-	ErrTransactionNotFoundForMerklePath = errors.New("transaction not found for given Merkle path")
+	ErrMerklePathNotFoundForTransaction = errors.New("merkle path not found for transaction")
 )