+++ blocktx/store/sql/sql.go
@@ -12,7 +12,6 @@ import (
 	"github.com/labstack/gommon/random"
 	_ "github.com/lib/pq"
 	"github.com/ordishs/gocore"
-	"github.com/pkg/errors"
 	"github.com/spf13/viper"
 	_ "modernc.org/sqlite"
 )
@@ -61,32 +60,32 @@ func New(engine string) (*SQL, error) {
 
 		dbName := viper.GetString("blocktx.db.postgres.name")
 		if dbName == "" {
-			return nil, errors.Errorf("setting blocktx.db.postgres.name not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.name not found")
 		}
 
 		dbPassword := viper.GetString("blocktx.db.postgres.password")
 		if dbPassword == "" {
-			return nil, errors.Errorf("setting blocktx.db.postgres.password not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.password not found")
 		}
 
 		dbUser := viper.GetString("blocktx.db.postgres.user")
 		if dbUser == "" {
-			return nil, errors.Errorf("setting blocktx.db.postgres.user not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.user not found")
 		}
 
 		dbHost := viper.GetString("blocktx.db.postgres.host")
 		if dbHost == "" {
-			return nil, errors.Errorf("setting blocktx.db.postgres.host not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.host not found")
 		}
 
 		dbPort := viper.GetInt("blocktx.db.postgres.port")
 		if dbPort == 0 {
-			return nil, errors.Errorf("setting blocktx.db.postgres.port not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.port not found")
 		}
 
 		sslMode := viper.GetString("blocktx.db.postgres.sslMode")
 		if sslMode == "" {
-			return nil, errors.Errorf("setting blocktx.db.postgres.sslMode not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.sslMode not found")
 		}
 
 		dbInfo := fmt.Sprintf("user=%s password=%s dbname=%s host=%s port=%d sslmode=%s", dbUser, dbPassword, dbName, dbHost, dbPort, sslMode)
@@ -98,13 +97,13 @@ func New(engine string) (*SQL, error) {
 
 		idleConns := viper.GetInt("blocktx.db.postgres.maxIdleConns")
 		if idleConns == 0 {
-			return nil, errors.Errorf("setting blocktx.db.postgres.maxIdleConns not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.maxIdleConns not found")
 		}
 
 		db.SetMaxIdleConns(idleConns)
 		maxOpenConns := viper.GetInt("blocktx.db.postgres.maxOpenConns")
 		if maxOpenConns == 0 {
-			return nil, errors.Errorf("setting blocktx.db.postgres.maxOpenConns not found")
+			return nil, fmt.Errorf("setting blocktx.db.postgres.maxOpenConns not found")
 		}
 
 		db.SetMaxOpenConns(maxOpenConns)
@@ -119,7 +118,7 @@ func New(engine string) (*SQL, error) {
 		} else {
 			folder := viper.GetString("dataFolder")
 			if folder == "" {
-				return nil, errors.Errorf("setting dataFolder not found")
+				return nil, fmt.Errorf("setting dataFolder not found")
 			}
 			if err = os.MkdirAll(folder, 0o755); err != nil {
 				return nil, fmt.Errorf("failed to create data folder %s: %+v", folder, err)