+++ cmd/blocktx.go
@@ -2,23 +2,22 @@ package cmd
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"log/slog"
 	"os"
 	"time"
 
 	"github.com/bitcoin-sv/arc/blocktx"
+	"github.com/bitcoin-sv/arc/blocktx/async/nats_mq"
 	"github.com/bitcoin-sv/arc/config"
-	"github.com/spf13/viper"
 )
 
 const BecomePrimaryintervalSecs = 30
 
 func StartBlockTx(logger *slog.Logger) (func(), error) {
-	dbMode := viper.GetString("blocktx.db.mode")
-	if dbMode == "" {
-		return nil, errors.New("blocktx.db.mode not found in config")
+	dbMode, err := config.GetString("blocktx.db.mode")
+	if err != nil {
+		return nil, err
 	}
 
 	// dbMode can be sqlite, sqlite_memory or postgres
@@ -55,7 +54,38 @@ func StartBlockTx(logger *slog.Logger) (func(), error) {
 		return nil, err
 	}
 
-	peerHandler, err := blocktx.NewPeerHandler(logger, blockStore, startingBlockHeight, peerURLs, network, blocktx.WithRetentionDays(recordRetentionDays))
+	natsURL, err := config.GetString("queueURL")
+	if err != nil {
+		return nil, err
+	}
+
+	registerTxInterval, err := config.GetDuration("blocktx.registerTxsInterval")
+	if err != nil {
+		return nil, err
+	}
+
+	// The tx channel needs the capacity so that it could potentially buffer up to a certain nr of transactions per second
+	const targetTps = 6000
+	capacityRequired := int(registerTxInterval.Seconds() * targetTps)
+	if capacityRequired < 100 {
+		capacityRequired = 100
+	}
+
+	txChannel := make(chan []byte, capacityRequired)
+	consumer, err := nats_mq.NewNatsMQConsumer(txChannel, logger, natsURL)
+	if err != nil {
+		return nil, err
+	}
+
+	err = consumer.ConsumeTransactions()
+	if err != nil {
+		return nil, err
+	}
+
+	peerHandler, err := blocktx.NewPeerHandler(logger, blockStore, startingBlockHeight, peerURLs, network,
+		blocktx.WithRetentionDays(recordRetentionDays),
+		blocktx.WithTxChan(txChannel),
+		blocktx.WithRegisterTxsInterval(registerTxInterval))
 	if err != nil {
 		return nil, err
 	}
@@ -87,6 +117,12 @@ func StartBlockTx(logger *slog.Logger) (func(), error) {
 
 	return func() {
 		logger.Info("Shutting down blocktx store")
+
+		err = consumer.Shutdown()
+		if err != nil {
+			logger.Error("failed to shutdown consumer", slog.String("err", err.Error()))
+		}
+
 		err = blockStore.Close()
 		if err != nil {
 			logger.Error("Error closing blocktx store", slog.String("err", err.Error()))