+++ k8s_watcher/watcher.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/bitcoin-sv/arc/metamorph"
 	"github.com/bitcoin-sv/arc/metamorph/metamorph_api"
 )
 
@@ -21,7 +22,7 @@ type K8sClient interface {
 }
 
 type Watcher struct {
-	metamorphClient  metamorph_api.MetaMorphAPIClient
+	metamorphClient  metamorph.TransactionMaintainer
 	k8sClient        K8sClient
 	logger           *slog.Logger
 	ticker           Ticker
@@ -39,7 +40,7 @@ func WithLogger(logger *slog.Logger) func(*Watcher) {
 type ServerOption func(f *Watcher)
 
 // New The K8s watcher listens to events coming from Kubernetes. If it detects a metamorph pod which was terminated, then it sets records locked by this pod to unlocked. This is a safety measure for the case that metamorph is terminated ungracefully where it misses to unlock its records itself.
-func New(client metamorph_api.MetaMorphAPIClient, k8sClient K8sClient, namespace string, opts ...ServerOption) *Watcher {
+func New(client metamorph.TransactionMaintainer, k8sClient K8sClient, namespace string, opts ...ServerOption) *Watcher {
 	watcher := &Watcher{
 		metamorphClient: client,
 		k8sClient:       k8sClient,
@@ -114,7 +115,7 @@ func (c *Watcher) Start() error {
 							continue
 						}
 
-						c.logger.Info("records unlocked", slog.Int("rows-affected", int(resp.RecordsAffected)), slog.String("pod-name", podName))
+						c.logger.Info("records unlocked", slog.Int64("rows-affected", resp), slog.String("pod-name", podName))
 					}
 				}