+++ metamorph/async/publisher.go
@@ -1,0 +1,5 @@
+package async
+
+type Publisher interface {
+	PublishTransaction(hash []byte) error
+}