+++ metamorph/server_test.go
@@ -230,7 +230,7 @@ func TestServer_GetTransactionStatus(t *testing.T) {
 
 			want: nil,
 			wantErr: func(t assert.TestingT, err error, rest ...interface{}) bool {
-				return assert.ErrorIs(t, err, store.ErrNotFound, rest...)
+				return assert.ErrorIs(t, err, ErrNotFound, rest...)
 			},
 		},
 		{
@@ -275,7 +275,7 @@ func TestServer_GetTransactionStatus(t *testing.T) {
 				Txid: testdata.TX1,
 			},
 			status:             metamorph_api.Status_MINED,
-			getTxMerklePathErr: blocktx.ErrTransactionNotFoundForMerklePath,
+			getTxMerklePathErr: blocktx.ErrMerklePathNotFoundForTransaction,
 
 			want: &metamorph_api.TransactionStatus{
 				StoredAt:    timestamppb.New(testdata.Time),